# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    AgentRole:
      type: string
      enum: [ROUTA, CRAFTER, GATE, DEVELOPER]
...
    TaskStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, REVIEW_REQUIRED, COMPLETED, NEEDS_FIX, BLOCKED, CANCELLED]
...
    VerificationVerdict:
      type: string
      enum: [APPROVED, NOT_APPROVED, BLOCKED]
...
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    Task:
      type: object
      required: [id, title, objective, status, dependencies, workspaceId, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        objective:
          type: string
        scope:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        verificationCommands:
          type: array
          items:
            type: string
        assignedTo:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        boardId:
          type: string
        columnId:
          type: string
        position:
          type: integer
        priority:
          $ref: "#/components/schemas/TaskPriority"
        labels:
          type: array
          items:
            type: string
        assignee:
          type: string
        assignedProvider:
          type: string
        assignedRole:
          $ref: "#/components/schemas/AgentRole"
        assignedSpecialistId:
          type: string
        assignedSpecialistName:
          type: string
        triggerSessionId:
          type: string
        githubId:
          type: string
        githubNumber:
          type: integer
        githubUrl:
          type: string
        githubRepo:
          type: string
        githubState:
          type: string
        githubSyncedAt:
          type: string
          format: date-time
        lastSyncError:
          type: string
        dependencies:
          type: array
          items:
            type: string
        parallelGroup:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completionSummary:
          type: string
        verificationVerdict:
          $ref: "#/components/schemas/VerificationVerdict"
        verificationReport:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          const: true

    # ── Sandbox schemas ──
...
    Worktree:
      type: object
      required: [id, codebaseId, workspaceId, worktreePath, branch, baseBranch, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        codebaseId:
          type: string
        workspaceId:
          type: string
        worktreePath:
          type: string
        branch:
          type: string
        baseBranch:
          type: string
        status:
          type: string
          enum: [creating, active, error, removing]
        sessionId:
          type: string
        label:
          type: string
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

# ─────────────────────────────────────────────────────────
# API Paths
# ─────────────────────────────────────────────────────────
paths:
...
  /api/tasks/{id}:
    get:
      operationId: getTask
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                objective:
                  type: string
                status:
                  $ref: "#/components/schemas/TaskStatus"
                boardId:
                  type: string
                columnId:
                  type: string
                position:
                  type: integer
                priority:
                  $ref: "#/components/schemas/TaskPriority"
                labels:
                  type: array
                  items:
                    type: string
                assignee:
                  type: string
                assignedProvider:
                  type: string
                assignedRole:
                  $ref: "#/components/schemas/AgentRole"
                assignedSpecialistId:
                  type: string
                assignedSpecialistName:
                  type: string
                repoPath:
                  type: string
                syncToGitHub:
                  type: boolean
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...
  /api/workspaces/{workspace_id}/codebases/{codebase_id}/worktrees:
    get:
      operationId: listWorktrees
      summary: List worktrees for a codebase
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: codebase_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of worktrees
          content:
            application/json:
              schema:
                type: object
                properties:
                  worktrees:
                    type: array
                    items:
                      $ref: "#/components/schemas/Worktree"
    post:
      operationId: createWorktree
      summary: Create a new git worktree
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: codebase_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                branch:
                  type: string
                baseBranch:
                  type: string
                label:
                  type: string
      responses:
        "201":
          description: Created worktree
          content:
            application/json:
              schema:
                type: object
                properties:
                  worktree:
                    $ref: "#/components/schemas/Worktree"
        "404":
          description: Codebase or workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Branch already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Git operation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...